factor out carla_sim logic
